api/requirements.txt
fastapi
uvicorn[standard]
pandas
rapidfuzz
python-dotenv
